bootstrap link fix and jumbotron work
fixed the issue that the link for the bootstrap comes before the link
for the stylesheet for all the other pages (was only done for
index.html).
Introduced an id for the first jumbotron on the index side and gave it a
top margin of 100px so it no longer sticks to the navbar.
Introduced "pictures" of us on the about page.
Created an Email formula and a link to social media on the Contact page.
Introduced Image to index page.
Designed the blog page with featurettes.
started creating a caroussel but have not been able to get it to work.